backend/requirements.txt
django
djangorestframework
markdown
django-filter
pillow
django-jsonfield
psycopg2-binary
djangorestframework-simplejwt